Port upstream test_tools.py into per-module test files; fix exposed regressions
The upstream test_tools.py (505 lines, 59 tests covering tools Kyle and
Rich added in PRs #41/#42 plus earlier upstream tools) was written
against the pre-refactor sync interface and fails under the modular
async code. Rather than delete those tests, relocate them into the
per-module split structure yonib05's refactor established:

- tests/test_accounts.py (new) — get_accounts, get_account_holdings,
  refresh_accounts
- tests/test_budgets.py (new) — get_budgets
- tests/test_financial.py (new) — get_cashflow
- tests/test_categories.py (extend) — Kyle's get_transaction_categories,
  get_transaction_category_groups, create_transaction_category
- tests/test_tags.py (extend) — Kyle's get_transaction_tags,
  create_transaction_tag, add_transaction_tag; plus fixture-style
  coverage of set_transaction_tags
- tests/test_transactions.py (extend) — Rich's enhanced get_transactions
  (new fields/filters, account_id->account_ids fix), the SDK-signature
  create_transaction/update_transaction (from PR #35), Kyle's
  categorize_transaction

The conftest.py autouse fixture now patches every tool module's
`get_monarch_client` import site, not just the origin module in
`monarch_mcp_server.client`. Without this, fixture-style tests that
don't manually @patch their tool module silently bypass the mock.

Porting surfaced two regressions in yonib05's refactor that this commit
also fixes:

1. tools/accounts.py: get_accounts dropped the `is_hidden` field that
   upstream's pre-refactor version returned.
2. tools/budgets.py: get_budgets reimplemented the buggy formatter that
   Rob deliberately replaced in #39 (the wrong response key produced
   empty results). Return the raw API response, like upstream.

uv.lock regeneration drops cp314 wheel entries that the upstream <3.14
Python constraint makes unreachable.
